+++ mepo.d/cmdline/parser.py
@@ -117,7 +117,7 @@ class MepoArgParser(object):
         status.add_argument(
             '--ignore-permissions',
             action = 'store_true',
-            help = 'Tells status command to ignore changes in file permissions.')
+            help = 'Tells command to ignore changes in file permissions.')
 
     def __restore_state(self):
         restore_state = self.subparsers.add_parser(
@@ -134,6 +134,14 @@ class MepoArgParser(object):
             '--name-only',
             action = 'store_true',
             help = 'Show only names of changed files')
+        diff.add_argument(
+            '--name-status',
+            action = 'store_true',
+            help = 'Show name-status of changed files')
+        diff.add_argument(
+            '--ignore-permissions',
+            action = 'store_true',
+            help = 'Tells command to ignore changes in file permissions.')
         diff.add_argument(
             '--staged',
             action = 'store_true',